#1819 trace format v2  - Refactoring SpanEncoder & SpanDecoder
