+++ commons-server/src/main/java/com/navercorp/pinpoint/common/server/bo/serializer/trace/v2/SpanSerializerV2.java
@@ -4,6 +4,7 @@ import com.navercorp.pinpoint.common.server.bo.SpanBo;
 import com.navercorp.pinpoint.common.server.bo.serializer.HbaseSerializer;
 import com.navercorp.pinpoint.common.server.bo.serializer.SerializationContext;
 import org.apache.hadoop.hbase.client.Put;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.nio.ByteBuffer;
@@ -18,7 +19,8 @@ import static com.navercorp.pinpoint.common.hbase.HBaseTables.TRACE_V2_CF_SPAN;
 public class SpanSerializerV2 implements HbaseSerializer<SpanBo, Put> {
 
 
-    private final SpanEncoder spanEncoder = new SpanEncoder();
+    @Autowired
+    private SpanEncoder spanEncoder;
 
     public SpanSerializerV2() {
     }
